Fix RenderSystem layer data type error

+++ src/GameEngine/Rendering/RenderSystem.cpp
@@ -6,7 +6,7 @@ RenderSystem::RenderSystem(const Game* game, Sprite* canvas) : GameSystem(game),
 
 void RenderSystem::update(uint32_t deltaMicros){
 	std::set<int16_t> layers;
-	std::map<uint8_t, std::vector<std::shared_ptr<const GameObject>>> layerMap;
+	std::map<int16_t, std::vector<std::shared_ptr<const GameObject>>> layerMap;
 
 
 	//map objects to layers
